+++ Example/libraries/zf_driver/zf_driver_pwm.h
@@ -50,6 +50,6 @@ struct pwm_info
 };
 
 void pwm_get_dev_info(const char *path, struct pwm_info *pwm_info);
-void pwm_set_duty(const char *path, uint16 duty);
+void pwm_set_duty(const char *path, uint32 duty);
 
 #endif
